+++ collections/settings.js
@@ -294,6 +294,15 @@ settingsSchemaObject = {
       private: true
     }
   },
+  debug: {
+    type: Boolean,
+    optional: true,
+    label: 'Debug Mode',
+    autoform: {
+      group: 'debug',
+      instructions: 'Enable debug mode for more details console logs'
+    }
+  }
 };
 
 // add any extra properties to settingsSchemaObject (provided by packages for example)
@@ -329,3 +338,9 @@ if (Meteor.isClient){
     }
   });
 }
+
+Meteor.startup(function () {
+  // override Meteor.absoluteUrl() with URL provided in settings
+  Meteor.absoluteUrl.defaultOptions.rootUrl = getSetting('siteUrl', Meteor.absoluteUrl());
+  debug = getSetting('debug', false);
+});